# openapi.yaml
openapi: 3.1.0
info:
  title: Anamnesis HTTP API
  version: 1.0.0
  summary: Narrow clinician-in-the-loop workflow API for case intake, bounded evidence ingestion, packet drafting, review, and audit.
  description: |
    The Anamnesis API is a narrow workflow surface for organizing intake, source artifacts,
    bounded text and FHIR import seams, physician packet drafts, explicit clinician review,
    packet finalization, and append-only audit history.

    This API is not a diagnosis, treatment, prescription, or general-purpose FHIR server API.

    Authentication model:
    - application endpoints expect `Authorization: Bearer <API_KEY>` when `API_KEY` is configured;
    - application endpoints also accept `Authorization: Bearer <JWT>` when `JWT_SECRET` or `JWT_PUBLIC_KEY` is configured;
    - `JWT_SECRET` and `JWT_PUBLIC_KEY` are mutually exclusive JWT verifier modes;
    - when `API_KEY` and one JWT verifier mode are configured, either bearer mechanism may be used;
    - JWTs must carry a non-empty `sub`, must satisfy configured `iss`/`aud` constraints, and may additionally be required to match a configured JOSE `typ` value;
    - `JWT_PUBLIC_KEY` enables RS256 verification and requires an RSA public key with modulus length >= 2048 bits;
    - unauthenticated startup is allowed only when `ALLOW_INSECURE_DEV_AUTH=true` is explicitly set;
    - the insecure development override is rejected when `NODE_ENV=production`.

    Interoperability boundary:
...
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
  - jwtBearer: []
paths:
  /api/cases:
    post:
      tags: [Cases]
      summary: Create case
      operationId: createCase
      description: Create a case with structured intake and an optional workflow family. New cases start in `INTAKING`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseRequest'
      responses:
        '201':
          description: Case created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
    get:
      tags: [Cases]
      summary: List cases
      operationId: listCases
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Cases listed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /api/cases/{caseId}:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    get:
      tags: [Cases]
      summary: Get case
      operationId: getCase
      responses:
        '200':
          description: Case retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
    delete:
      tags: [Cases]
      summary: Delete case
      operationId: deleteCase
      description: Delete a case and append a `case.deleted` audit event.
      responses:
        '204':
          description: Case deleted.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /api/cases/{caseId}/evidence-lineage:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    get:
      tags: [Lineage]
      summary: Get evidence lineage graph
      operationId: getEvidenceLineage
      description: Return a read-only producer-to-consumer artifact graph plus artifact summary metadata for one case.
      responses:
        '200':
          description: Evidence lineage graph returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceLineageResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /api/cases/{caseId}/artifacts:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Artifacts]
      summary: Add source artifact
      operationId: addArtifact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddArtifactRequest'
      responses:
        '201':
          description: Artifact added and case returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /api/cases/{caseId}/artifacts/{artifactId}:
    parameters:
      - $ref: '#/components/parameters/CaseId'
      - $ref: '#/components/parameters/ArtifactId'
    delete:
      tags: [Artifacts]
      summary: Remove source artifact
      operationId: removeArtifact
      responses:
        '200':
          description: Artifact removed and case returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /api/cases/{caseId}/samples:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Extraction]
      summary: Register sample
      operationId: registerSample
      description: Register a case-scoped molecular sample for downstream review workflows.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterSampleRequest'
      responses:
        '201':
          description: Sample registered and case returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /api/cases/{caseId}/study-context:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Extraction]
      summary: Attach study context
      operationId: attachStudyContext
      description: Attach imaging study metadata and initialize pending QC state for case-scoped review workflows.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachStudyContextRequest'
      responses:
        '200':
          description: Study context attached and case returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /api/cases/{caseId}/qc-summary:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Extraction]
      summary: Record QC summary
      operationId: recordQcSummary
      description: Record a QC disposition and optional QC detail for an attached imaging study context.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordQcSummaryRequest'
      responses:
        '200':
          description: QC summary recorded and case returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
...
  /api/cases/{caseId}/audit-events:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    get:
      tags: [Audit]
      summary: List audit events by case id
      operationId: listAuditEventsByCase
      description: |
        Return append-only audit history by `caseId`. This endpoint is keyed to the audit store and may still
        return events after the case itself has been deleted.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Audit events returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque API key
    jwtBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token validated via HS256 when `JWT_SECRET` is configured or via RS256 when `JWT_PUBLIC_KEY` is configured. Runtime validation enforces signature, `exp`, `iat`, `nbf`, non-empty `sub`, configured `iss`/`aud`, optional JOSE `typ`, and an RSA-2048 minimum for RS256 verification keys.
  parameters:
    CaseId:
      name: caseId
      in: path
      required: true
      description: Case UUID.
      schema:
        type: string
        format: uuid
    ArtifactId:
      name: artifactId
      in: path
      required: true
      description: Artifact UUID.
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return (1-1000, default 100).
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    Offset:
      name: offset
      in: query
      required: false
      description: Number of items to skip (default 0).
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
...
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    UnauthorizedError:
      description: Missing or invalid bearer token.
      headers:
        WWW-Authenticate:
          description: Bearer authentication challenge.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFoundError:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Workflow or domain conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequestsError:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    DateOnly:
      type: string
      pattern: '^\d{4}-\d{2}-\d{2}$'
      examples: ['2026-04-03']
    CaseStatus:
      type: string
      enum: [INTAKING, READY_FOR_PACKET, REVIEW_REQUIRED]
    WorkflowFamily:
      type: string
      enum: [GENERAL_INTAKE, MRI_SECOND_OPINION, MRNA_BOARD_REVIEW]
    ArtifactType:
      type: string
      enum: [note, lab, summary, report, imaging-summary]
    ArtifactClass:
      type: string
      enum: [SOURCE, DERIVED, REVIEW_PACKET, HANDOFF_PACKET]
    SourceArtifactSemanticType:
      type: string
      enum:
        [clinical-note, lab-panel, clinical-summary, imaging-study, imaging-qc-summary, tumor-dna-fastq, normal-dna-fastq, tumor-rna-fastq, board-evidence-bundle]
...
    SampleType:
      type: string
      enum: [TUMOR_DNA, NORMAL_DNA, TUMOR_RNA, FOLLOW_UP]
    AssayType:
      type: string
      enum: [WES, WGS, RNA_SEQ, PANEL, OTHER]
    QcCheckStatus:
      type: string
      enum: [pass, warn, reject]
...
    PaginationMeta:
      type: object
      additionalProperties: false
      required: [returnedCount, limit, offset]
      properties:
        returnedCount:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
          maximum: 1000
        offset:
          type: integer
          minimum: 0
    CaseIntake:
      type: object
      additionalProperties: false
      required: [chiefConcern, symptomSummary, historySummary, questionsForClinician]
      properties:
        chiefConcern:
          type: string
          minLength: 1
          maxLength: 200
        symptomSummary:
          type: string
          minLength: 1
          maxLength: 4000
        historySummary:
          type: string
          minLength: 1
          maxLength: 4000
        questionsForClinician:
          type: array
          maxItems: 12
          items:
            type: string
            minLength: 1
            maxLength: 300
...
    EvidenceLineageGraph:
      type: object
      additionalProperties: false
      required: [edges, roots, terminal]
      properties:
        edges:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceLineageEdge'
        roots:
          type: array
          items:
            type: string
            format: uuid
        terminal:
          type: array
          items:
            type: string
            format: uuid
    EvidenceLineageArtifactSummary:
      type: object
      additionalProperties: false
      required: [artifactId, artifactType, title, summary]
      properties:
        artifactId:
          type: string
          format: uuid
        artifactType:
          $ref: '#/components/schemas/ArtifactType'
        artifactClass:
          $ref: '#/components/schemas/ArtifactClass'
        semanticType:
          $ref: '#/components/schemas/SourceArtifactSemanticType'
        title:
          type: string
        summary:
          type: string
        derivedFromArtifactIds:
          type: array
          items:
            type: string
            format: uuid
        provenance:
          type: string
    EvidenceLineageMeta:
      type: object
      additionalProperties: false
      required: [artifactCount, edgeCount]
      properties:
        artifactCount:
          type: integer
          minimum: 0
        edgeCount:
          type: integer
          minimum: 0
    EvidenceLineageResponse:
      type: object
      additionalProperties: false
      required: [lineage, artifacts, meta]
      properties:
        lineage:
          $ref: '#/components/schemas/EvidenceLineageGraph'
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceLineageArtifactSummary'
        meta:
          $ref: '#/components/schemas/EvidenceLineageMeta'
    SampleRecord:
      type: object
      additionalProperties: false
      required: [sampleId, sampleType, assayType, accessionId, sourceSite, registeredAt]
      properties:
        sampleId:
          type: string
        sampleType:
          $ref: '#/components/schemas/SampleType'
        assayType:
          $ref: '#/components/schemas/AssayType'
        accessionId:
          type: string
        sourceSite:
          type: string
        registeredAt:
          type: string
          format: date-time
...
    StudyContextRecord:
      type: object
      additionalProperties: false
      required:
        [studyInstanceUid, dicomStudyInstanceUid, accessionNumber, studyDate, sourceArchive, dicomWebBaseUrl, metadataSummary, series, receivedAt, source]
      properties:
        studyInstanceUid:
          type: string
        dicomStudyInstanceUid:
          type: string
        accessionNumber:
          oneOf:
            - type: string
            - type: 'null'
        studyDate:
          oneOf:
            - $ref: '#/components/schemas/DateOnly'
            - type: 'null'
        sourceArchive:
          oneOf:
            - type: string
            - type: 'null'
        dicomWebBaseUrl:
          oneOf:
            - type: string
              format: uri
            - type: 'null'
        metadataSummary:
          type: array
          items:
            type: string
        series:
          type: array
          items:
            $ref: '#/components/schemas/StudySeriesRecord'
        receivedAt:
          type: string
          format: date-time
        source:
          type: string
          enum: [public-api, internal-ingest]
...
    AnamnesisCase:
      type: object
      additionalProperties: false
      required:
        - caseId
        - workflowFamily
        - status
        - createdAt
        - updatedAt
        - intake
        - samples
        - artifacts
        - physicianPackets
      properties:
        caseId:
          type: string
          format: uuid
        patientLabel:
          type: string
        workflowFamily:
          $ref: '#/components/schemas/WorkflowFamily'
        status:
          $ref: '#/components/schemas/CaseStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        intake:
          $ref: '#/components/schemas/CaseIntake'
        samples:
          type: array
          items:
            $ref: '#/components/schemas/SampleRecord'
        studyContext:
          $ref: '#/components/schemas/StudyContextRecord'
        qcSummary:
          $ref: '#/components/schemas/QcSummaryRecord'
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/SourceArtifact'
        physicianPackets:
          type: array
          items:
            $ref: '#/components/schemas/PhysicianPacket'
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    ValidationErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
          const: invalid_input
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
    CaseEnvelope:
      type: object
      additionalProperties: false
      required: [case]
      properties:
        case:
          $ref: '#/components/schemas/AnamnesisCase'
    CaseListResponse:
      type: object
      additionalProperties: false
      required: [cases, meta]
      properties:
        cases:
          type: array
          items:
            $ref: '#/components/schemas/AnamnesisCase'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    AuditEventListResponse:
      type: object
      additionalProperties: false
      required: [events, meta]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/AuditEventRecord'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    CreateCaseRequest:
      type: object
      additionalProperties: false
      required: [intake]
      properties:
        patientLabel:
          type: string
          minLength: 1
          maxLength: 120
        workflowFamily:
          $ref: '#/components/schemas/WorkflowFamily'
        intake:
          $ref: '#/components/schemas/CaseIntake'
    AddArtifactRequest:
      type: object
      additionalProperties: false
      required: [artifactType, title, summary]
      properties:
        artifactType:
          $ref: '#/components/schemas/ArtifactType'
        artifactClass:
          $ref: '#/components/schemas/ArtifactClass'
        semanticType:
          $ref: '#/components/schemas/SourceArtifactSemanticType'
        sampleId:
          type: string
          minLength: 1
          maxLength: 120
        artifactHash:
          type: string
          minLength: 1
          maxLength: 300
        storageUri:
          type: string
          minLength: 1
          maxLength: 1000
        mediaType:
          type: string
          minLength: 1
          maxLength: 160
        derivedFromArtifactIds:
          type: array
          minItems: 1
          maxItems: 64
          items:
            type: string
            format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        summary:
          type: string
          minLength: 1
          maxLength: 4000
        sourceDate:
          $ref: '#/components/schemas/DateOnly'
        provenance:
          type: string
          minLength: 1
          maxLength: 300
    RegisterSampleRequest:
      type: object
      additionalProperties: false
      required: [sampleId, sampleType, assayType, accessionId, sourceSite]
      properties:
        sampleId:
          type: string
          minLength: 1
          maxLength: 120
        sampleType:
          $ref: '#/components/schemas/SampleType'
        assayType:
          $ref: '#/components/schemas/AssayType'
        accessionId:
          type: string
          minLength: 1
          maxLength: 120
        sourceSite:
          type: string
          minLength: 1
          maxLength: 160
...
    StudyContextInput:
      type: object
      additionalProperties: false
      properties:
        studyInstanceUid:
          type: string
          minLength: 1
          maxLength: 200
        accessionNumber:
          type: string
          minLength: 1
          maxLength: 120
        studyDate:
          $ref: '#/components/schemas/DateOnly'
        sourceArchive:
          type: string
          minLength: 1
          maxLength: 160
        dicomWebBaseUrl:
          type: string
          format: uri
          maxLength: 1000
        metadataSummary:
          type: array
          maxItems: 32
          items:
            type: string
            minLength: 1
            maxLength: 300
        series:
          type: array
          maxItems: 64
          items:
            $ref: '#/components/schemas/StudySeriesInput'
    AttachStudyContextRequest:
      type: object
      additionalProperties: false
      required: [source]
      properties:
        source:
          type: string
          enum: [public-api, internal-ingest]
        studyContext:
          $ref: '#/components/schemas/StudyContextInput'
    QcMetricInput:
      type: object
      additionalProperties: false
      required: [name, value]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        value:
          type: number
        unit:
          type: string
          minLength: 1
          maxLength: 32
    QcCheckInput:
      type: object
      additionalProperties: false
      required: [checkId, status, detail]
      properties:
        checkId:
          type: string
          minLength: 1
          maxLength: 120
        status:
          $ref: '#/components/schemas/QcCheckStatus'
        detail:
          type: string
          minLength: 1
          maxLength: 500
    RecordQcSummaryRequest:
      type: object
      additionalProperties: false
      required: [disposition]
      properties:
        disposition:
          $ref: '#/components/schemas/QcCheckStatus'
        issues:
          type: array
          maxItems: 32
          items:
            type: string
            minLength: 1
            maxLength: 300
        qcSummary:
          type: object
          additionalProperties: false
          properties:
            summary:
              type: string
              minLength: 1
              maxLength: 4000
            checks:
              type: array
              maxItems: 64
              items:
                $ref: '#/components/schemas/QcCheckInput'
            metrics:
              type: array
              maxItems: 64
              items:
                $ref: '#/components/schemas/QcMetricInput'
...